requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
astunparse==1.6.3
attrs==23.2.0
braceexpand==0.1.7
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
discord.py==2.3.2
frozenlist==1.4.1
idna==3.6
import-expression==1.1.4
jishaku==2.5.2
multidict==6.0.5
python-dotenv==1.0.1
requests==2.31.0
six==1.16.0
urllib3==2.2.1
yarl==1.9.4